+++ jarvis/modules/microphone/graph_mic.py
@@ -192,6 +192,12 @@ def _kick_off() -> NoReturn:
 
 
 if __name__ == '__main__':
+    # imports within __main__ to avoid potential/future path error and circular import
+    # override 'current_process().name' to avoid being set as 'MainProcess'
+    # importing at top level requires setting current_process().name at top level which will in turn override any import
+    from multiprocessing import current_process
+
+    current_process().name = "PlotMic"
     from jarvis.modules.database import database
     from jarvis.modules.logger import config
     from jarvis.modules.logger.custom_logger import logger